feat: Modularize block checking logic
This patch splits the checker into multiple files, each responsible for
checking different aspects of the language; it then converts the block
checking logic from repetitive functions into a schema-based approach.
See `checker/blocks.ts` for the new logic. This replaces functions like
`checkTrack`, `checkMixer`, `checkBus`, `checkInstrument`, etc.
