+++ packages/language/src/compiler/checker/capabilities.ts
@@ -1,0 +1,9 @@
+import type { Capabilities } from '../../type-system/base/function.ts'
+
+export const noCapabilities: Capabilities = { mayBlock: false }
+
+export function mergeCapabilities (target: Capabilities, source: Capabilities): Capabilities {
+  return {
+    mayBlock: target.mayBlock || source.mayBlock
+  }
+}
